Rename startEvent shape to more generic "eventShape"
