+++ src/components/nodes/startEvent/eventShape.js
